+++ ts/CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add output schema option
+
 ## [0.7.0] — 2026-05-07
 
 ### Added
@@ -56,6 +62,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0] — 2026-05-02
 
+[unreleased]: https://github.com/mantyx-io/mantyx-sdk/compare/v0.7.0..HEAD
 [0.7.0]: https://github.com/mantyx-io/mantyx-sdk/compare/v0.6.0..v0.7.0
 [0.6.0]: https://github.com/mantyx-io/mantyx-sdk/compare/v0.5.0..v0.6.0
 [0.5.0]: https://github.com/mantyx-io/mantyx-sdk/compare/v0.4.1..v0.5.0